requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.1
django-markupfield==2.0.1
django-model-utils==4.2.0
django-taggit==3.0.0
gunicorn==20.1.0
Pillow==9.2.0
psycopg2-binary==2.9.3
sqlparse==0.4.2
whitenoise==6.2.0
